requirements.txt
-i https://mirrors.aliyun.com/pypi/simple
astroid==2.2.5
autopep8==1.4.4
django-mdeditor==0.1.16
django==2.2.4
isort==4.3.21
lazy-object-proxy==1.4.1
markdown2==2.3.8
markdown==3.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
yapf==0.28.0
